build: move strip into its own step
Within the Makefile, mame is first linked together and then stripped. If
the strip step fails, the end result is an unstripped mame binary.
Running 'make' again will not try to re-strip it because a file named
'mame' now exists.

Modify the final steps of the build process. Place the unstripped binary
into $(OBJ), and the stripped binary into $(PWD) as before.